refactor: モデルフォルダ内を細かく整理
