+++ lib/models/request/weather_condition_request.freezed.dart
