+++ lib/models/response/weather_condition_response.dart
