+++ lib/ui/screens/home_screen.dart
@@ -3,8 +3,8 @@ import 'dart:convert';
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:flutter_training/models/weather_condition.dart';
-import 'package:flutter_training/models/weather_condition_request.dart';
-import 'package:flutter_training/models/weather_condition_response.dart';
+import 'package:flutter_training/models/request/weather_condition_request.dart';
+import 'package:flutter_training/models/response/weather_condition_response.dart';
 import 'package:flutter_training/ui/extensions/weather_condition_ext.dart';
 import 'package:flutter_training/ui/widgets/button_row.dart';
 import 'package:flutter_training/ui/widgets/temperature_row.dart';
